chore: move space components into a folder
